refactor: move types in `resolve` module to seperate files
It was getting hard to find the functions that you needed while
browsing the module. So moved some types to their own files.
No functionality has changed
