+++ crates/rattler_installs_packages/tests/resolver.rs
@@ -1,13 +1,12 @@
 #![cfg(feature = "resolvo")]
 
 use pep508_rs::{MarkerEnvironment, Requirement};
+use rattler_installs_packages::resolve::solve_options::{ResolveOptions, SDistResolution};
 use rattler_installs_packages::{
     index::PackageDb,
     python_env::{WheelTag, WheelTags},
     resolve::resolve,
     resolve::PinnedPackage,
-    resolve::ResolveOptions,
-    resolve::SDistResolution,
     types::NormalizedPackageName,
 };
 use std::{collections::HashMap, path::Path, str::FromStr, sync::OnceLock};
